No more warnings and some optimizations
