+++ examples/autobahn-client.rs
@@ -4,6 +4,7 @@ extern crate rustc_serialize as serialize;
 use std::str::from_utf8;
 use websocket::client::request::Url;
 use websocket::{Client, Message, Sender, Receiver};
+use websocket::message::Type;
 use websocket::dataframe::Opcode;
 use serialize::json;
 
@@ -48,18 +49,18 @@ fn main() {
 			};
 
 			match message.opcode {
-				Opcode::Text => {
+				Type::Text => {
                     let response = Message::text(from_utf8(&*message.payload).unwrap());
 					sender.send_message(&response).unwrap();
 				}
-				Opcode::Binary => {
+				Type::Binary => {
 					sender.send_message(&Message::binary(message.payload)).unwrap();
 				}
-				Opcode::Close => {
+				Type::Close => {
 					let _ = sender.send_message(&Message::close());
 					break;
 				}
-				Opcode::Ping => {
+				Type::Ping => {
 					sender.send_message(&Message::pong(message.payload)).unwrap();
 				}
 				_ => (),
@@ -98,15 +99,15 @@ fn get_case_count(addr: String) -> usize {
 			}
 		};
 		match message.opcode {
-			Opcode::Text => {
+			Type::Text => {
 				count = json::decode(from_utf8(&*message.payload).unwrap()).unwrap();
 				println!("Will run {} cases...", count);
 			}
-			Opcode::Close => {
+			Type::Close => {
 				let _ = sender.send_message(&Message::close());
 				break;
 			}
-			Opcode::Ping => {
+			Type::Ping => {
 				sender.send_message(&Message::pong(message.payload)).unwrap();
 			}
 			_ => (),
@@ -142,13 +143,13 @@ fn update_reports(addr: String, agent: &str) {
 			}
 		};
 		match message.opcode {
-			Opcode::Close => {
+			Type::Close => {
 				let _ = sender.send_message(&Message::close());
 				println!("Reports updated.");
 				println!("Test suite finished!");
 				return;
 			}
-			Opcode::Ping => {
+			Type::Ping => {
 				sender.send_message(&Message::pong(message.payload)).unwrap();
 			}
 			_ => (),
